# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam project",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token should manually have been added as a cookie by default but if that does not work the JWT token can be added here with the prefix: Bearer followed by the token"
    }
  },
  "paths": {
...
    "/admin/products": {
      "get": {
        "tags": [
          "Admin Front End"
        ],
        "description": "Admin products page",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin Front End"
        ],
        "description": "Admin orders page",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get all orders for a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/all": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get all orders (Admin only)",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/status/{orderId}": {
      "put": {
        "tags": [
          "Order"
        ],
        "description": "Update the status of an order (Admin only)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated status of the order.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get all products",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Creates a new product (Admin only)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Product details for new product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get a single product",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "Updating a product (Admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated details of the product.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "Deleting a product (Admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Product ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/membership/": {
      "get": {
        "tags": [
          "Membership"
        ],
        "description": "Get all membership statuses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Membership"
        ],
        "description": "Creates a new membership status (Admin only)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Membership status details for new status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Membership"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/membership/{id}": {
      "put": {
        "tags": [
          "Membership"
        ],
        "description": "Updating a membership status (Admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated details of the membership status.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateMembership"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Membership"
        ],
        "description": "Deleting a membership status (Admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Membership ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Playstation 5"
        },
        "description": {
          "type": "string",
          "example": "Sony's latest gaming console"
        },
        "price": {
          "type": "number",
          "example": 699
        },
        "quantity": {
          "type": "number",
          "example": 10
        },
        "discount": {
          "type": "number",
          "example": 0
        },
        "categoryId": {
          "type": "number",
          "example": 1
        },
        "brandId": {
          "type": "number",
          "example": 1
        },
        "isDeleted": {
          "type": "boolean",
          "example": false
        },
        "imgURL": {
          "type": "string",
          "example": "https://gmedia.playstation.com/is/image/SIEPDC/ps5-product-thumbnail-01-en-14sep21"
        }
      },
      "required": [
        "name",
        "description",
        "price",
        "quantity",
        "discount",
        "categoryId",
        "brandId",
        "isDeleted",
        "imgURL"
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Playstation 6"
        },
        "description": {
          "type": "string",
          "example": "Sony's futuristic gaming console"
        },
        "price": {
          "type": "number",
          "example": 8000
        },
        "quantity": {
          "type": "number",
          "example": 0
        },
        "discount": {
          "type": "number",
          "example": 0
        },
        "categoryId": {
          "type": "number",
          "example": 1
        },
        "brandId": {
          "type": "number",
          "example": 1
        },
        "isDeleted": {
          "type": "boolean",
          "example": false
        },
        "imgURL": {
          "type": "string",
          "example": "https://i.ytimg.com/vi/QNwJdTqZvts/maxresdefault.jpg"
        }
      },
      "required": [
        "name",
        "description",
        "price",
        "quantity",
        "discount",
        "categoryId",
        "brandId",
        "isDeleted",
        "imgURL"
      ]
    },
    "Membership": {
      "type": "object",
      "properties": {
        "statusName": {
          "type": "string",
          "example": "Platinum"
        },
        "discount": {
          "type": "number",
          "example": 50
        }
      },
      "required": [
        "statusName",
        "discount"
      ]
    },
    "UpdateMembership": {
      "type": "object",
      "properties": {
        "statusName": {
          "type": "string",
          "example": "Diamond"
        },
        "discount": {
          "type": "number",
          "example": 50
        }
      },
      "required": [
        "statusName",
        "discount"
      ]
    },
...
    "Order": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "Ordered"
        }
      },
      "required": [
        "status"
      ]
    }
  }
}